backend/requirements.txt
Django==5.1.2
django-summernote==0.8.20.0
djangorestframework==3.15.2